TeamEqupment Refactor
Refactored TeamEquipment
